Fixing infinite loop while calling collection_proxy#count
