+++ Gemfile
@@ -4,42 +4,43 @@ source 'https://rubygems.org'
 
 # Use this when testing with appraisal or Travis
 # ----------------------------------------------
-gemspec
+# gemspec
 
-gem 'rake'
-gem 'colored'
-gem 'pry'
-gem 'rspec'
-gem 'simplecov', require: false
-gem 'simplecov-summary'
+# gem 'rake'
+# gem 'colored'
+# gem 'pry'
+# gem 'rspec'
+# gem 'simplecov', require: false
+# gem 'simplecov-summary'
 
-gem 'appraisal', github: 'thoughtbot/appraisal'
+# gem 'appraisal', github: 'thoughtbot/appraisal'
 
-gem 'sqlite3', platform: :ruby
-gem 'jdbc-sqlite3', platform: :jruby, require: 'jdbc/sqlite3'
+# gem 'sqlite3', platform: :ruby
+# gem 'jdbc-sqlite3', platform: :jruby, require: 'jdbc/sqlite3'
 
-gem 'ruby-prof', platform: :ruby
+# gem 'ruby-prof', platform: :ruby
 
 
 
 # Use this when testing locally
 # ------------------------------------------
-# gem "colored"
-# gem "pry"
-# gem "rspec"
-# gem "simplecov", :require=>false
-# gem "simplecov-summary"
-# gem "appraisal", :github=>"thoughtbot/appraisal"
-# gem "sqlite3", :platform=>:ruby
-# gem "jdbc-sqlite3", :platform=>:jruby, :require=>"jdbc/sqlite3"
-# gem "ruby-prof", :platform=>:ruby
-# gem "activerecord", "~> 4.2.0", :require=>"active_record"
-# #gem "activerecord", "~> 4.1.13", :require=>"active_record"
-# gem "activerecord-jdbcsqlite3-adapter", :platform=>:jruby, :github=>"jruby/activerecord-jdbc-adapter"
-# gem "awesome_print"
-# gem 'pry'
-# gem 'pry-byebug'
-# gem 'pry-doc'
+gem "colored"
+gem "pry"
+gem "rspec"
+gem "simplecov", :require=>false
+gem "simplecov-summary"
+gem "appraisal", :github=>"thoughtbot/appraisal"
+gem "sqlite3", :platform=>:ruby
+gem "jdbc-sqlite3", :platform=>:jruby, :require=>"jdbc/sqlite3"
+gem "ruby-prof", :platform=>:ruby
+# gem 'activerecord', :git => 'https://github.com/rails/rails.git', tag: 'v5.1.2.rc1', :require=>"active_record"
+gem "activerecord", "~> 5.1.1", :require=>"active_record"
+#gem "activerecord", "~> 4.1.13", :require=>"active_record"
+gem "activerecord-jdbcsqlite3-adapter", :platform=>:jruby, :github=>"jruby/activerecord-jdbc-adapter"
+gem "awesome_print"
+gem 'pry'
+gem 'pry-byebug'
+gem 'pry-doc'
 
 
-# gemspec :path=>"./"
+gemspec :path=>"./"
