+++ lib/protector/adapters/active_record/collection_proxy.rb
@@ -6,9 +6,20 @@ module Protector
         extend ActiveSupport::Concern
         delegate :protector_subject, :protector_subject?, :to => :@association
 
+        def protector_relation
+          result = super
+          result.instance_variable_set(:@association, self.instance_variable_get(:@association).clone)
+          result
+        end
+
         def restrict!(*args)
           @association.restrict!(*args)
-          self
+          super
+        end
+
+        def unrestrict!
+          @association.unrestrict!
+          super
         end
       end
     end
