+++ lib/protector/adapters/active_record/relation.rb
@@ -67,22 +67,6 @@ module Protector
           super.restrict!(protector_subject)
         end
 
-        # @note This is here cause `NullRelation` can return `nil` from `count`
-        def count(*args)
-          super || 0
-        end
-
-        # @note This is here cause `NullRelation` can return `nil` from `sum`
-        def sum(*args)
-          super || 0
-        end
-
-        # Merges current relation with restriction and calls real `calculate`
-        def calculate(*args)
-          return super unless protector_subject?
-          protector_relation.unrestrict!.calculate(*args)
-        end
-
         # Merges current relation with restriction and calls real `exists?`
         def exists?(*args)
           return super unless protector_subject?
