+++ spec/spec_helpers/boot.rb
@@ -24,15 +24,15 @@ RSpec.configure do |config|
   config.run_all_when_everything_filtered = true
   config.filter_run :focus
 
-  config.after(:suite) do
-    if SimpleCov.running
-      silence_stream(STDOUT) do
-        SimpleCov::Formatter::HTMLFormatter.new.format(SimpleCov.result)
-      end
+  # config.after(:suite) do
+  #   # if SimpleCov.running
+  #   #   # silence_stream(STDOUT) do
+  #   #   #   SimpleCov::Formatter::HTMLFormatter.new.format(SimpleCov.result)
+  #   #   # end
 
-      SimpleCov::Formatter::SummaryFormatter.new.format(SimpleCov.result)
-    end
-  end
+  #   #   # SimpleCov::Formatter::SummaryFormatter.new.format(SimpleCov.result)
+  #   # end
+  # end
 
   # Run specs in random order to surface order dependencies. If you find an
   # order dependency and want to debug it, you can fix the order by providing
